# core/char: Speed up `to_digit()` for `radix <= 10`