--- a/src/libcore/char/methods.rs
+++ b/src/libcore/char/methods.rs
@@ -121,9 +121,7 @@ impl char {
     #[stable(feature = "rust1", since = "1.0.0")]
     #[inline]
     pub fn to_digit(self, radix: u32) -> Option<u32> {
-        if radix > 36 {
-            panic!("to_digit: radix is too high (maximum 36)");
-        }
+        assert!(radix <= 36, "to_digit: radix is too high (maximum 36)");
         let val = match self {
           '0' ..= '9' => self as u32 - '0' as u32,
           'a' ..= 'z' => self as u32 - 'a' as u32 + 10,