@scottmcm 2018-11-13
Consider whether there should also be a benchmark for non-constant radix.  I assume this will make that path slower, which is probably a good trade-off, but sounds worth quantifying.