@Turbo87 2018-11-13
@scottmcm yeah, I was wondering about that too. I'm not sure how to write such a benchmark though. I assume if I use the following code it would just inline the variable and use the constant radix path too:
```rust
let radix = 8u32;
c.to_digit(radix);
```

any clues on how I can make sure that the compiler does not consider this a constant? would `"32".parse::<u32>()` be an option?