@Xanewok 2018-11-13
@Turbo87 since we can use libtest here, maybe [`black_box`](https://doc.rust-lang.org/1.30.0/src/test/lib.rs.html#1539-1544) is worth a shot?