@Turbo87 2018-11-13
@Xanewok good idea. I assume that was just for blackboxing the output, but it seems it might also work for input. I'll try it out and report back 🤔