@Turbo87 2018-11-13
@scottmcm @Xanewok these are the results including the new `bench_to_digit_radix_var` benchmark:

### Before

```
# Run 1
test char::methods::bench_to_digit_radix_10        ... bench:      16,087 ns/iter (+/- 846)
test char::methods::bench_to_digit_radix_16        ... bench:      14,161 ns/iter (+/- 721)
test char::methods::bench_to_digit_radix_2         ... bench:      14,269 ns/iter (+/- 4,107)
test char::methods::bench_to_digit_radix_36        ... bench:      14,195 ns/iter (+/- 1,169)
test char::methods::bench_to_digit_radix_var       ... bench:      23,104 ns/iter (+/- 2,296)

# Run 2
test char::methods::bench_to_digit_radix_10        ... bench:      16,122 ns/iter (+/- 2,736)
test char::methods::bench_to_digit_radix_16        ... bench:      14,165 ns/iter (+/- 4,147)
test char::methods::bench_to_digit_radix_2         ... bench:      14,048 ns/iter (+/- 4,400)
test char::methods::bench_to_digit_radix_36        ... bench:      14,136 ns/iter (+/- 608)
test char::methods::bench_to_digit_radix_var       ... bench:      23,045 ns/iter (+/- 1,621)

# Run 3
test char::methods::bench_to_digit_radix_10        ... bench:      16,018 ns/iter (+/- 536)
test char::methods::bench_to_digit_radix_16        ... bench:      14,157 ns/iter (+/- 886)
test char::methods::bench_to_digit_radix_2         ... bench:      14,178 ns/iter (+/- 2,260)
test char::methods::bench_to_digit_radix_36        ... bench:      14,177 ns/iter (+/- 865)
test char::methods::bench_to_digit_radix_var       ... bench:      23,043 ns/iter (+/- 1,983)
```

### After

```
# Run 1
test char::methods::bench_to_digit_radix_10        ... bench:      11,518 ns/iter (+/- 497)
test char::methods::bench_to_digit_radix_16        ... bench:      14,623 ns/iter (+/- 1,260)
test char::methods::bench_to_digit_radix_2         ... bench:      11,240 ns/iter (+/- 1,430)
test char::methods::bench_to_digit_radix_36        ... bench:      14,189 ns/iter (+/- 854)
test char::methods::bench_to_digit_radix_var       ... bench:      24,513 ns/iter (+/- 3,337)

# Run 2
test char::methods::bench_to_digit_radix_10        ... bench:      11,536 ns/iter (+/- 1,549)
test char::methods::bench_to_digit_radix_16        ... bench:      14,602 ns/iter (+/- 1,033)
test char::methods::bench_to_digit_radix_2         ... bench:      13,940 ns/iter (+/- 9,252)
test char::methods::bench_to_digit_radix_36        ... bench:      14,303 ns/iter (+/- 749)
test char::methods::bench_to_digit_radix_var       ... bench:      24,298 ns/iter (+/- 1,440)

# Run 3
test char::methods::bench_to_digit_radix_10        ... bench:      11,491 ns/iter (+/- 840)
test char::methods::bench_to_digit_radix_16        ... bench:      14,540 ns/iter (+/- 991)
test char::methods::bench_to_digit_radix_2         ... bench:      11,275 ns/iter (+/- 576)
test char::methods::bench_to_digit_radix_36        ... bench:      14,201 ns/iter (+/- 444)
test char::methods::bench_to_digit_radix_var       ... bench:      24,617 ns/iter (+/- 5,132)
```

as predicted the variable radix case is slightly slower than before, but that seems like a good tradeoff to me as variable radix is probably much more rare than a constant radix.