@alexcrichton 2018-11-14
@bors: r+

These look like great improvement, thanks @Turbo87!